User_side/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
pymongo==4.6.3
dnspython>=2.4.0
bcrypt==4.1.2
python-dotenv==1.0.1
gunicorn
